Keyboard input almost done
Right now it changes the Active bool and renderer on cubes randomly
depending on the key clicked.  I'm working on something in the code
called boomBool, which is a bool on all of the cubes that sets them back
to white when triggered., as well as resets the active bool on the cube
to false.